Redesigned the interaction between the game and system components
Because native event handling is often part of the window manager,
and that it can receive window events that is only relevant to the
actual window, the system was redesigned, so event management and
graphics context is now in a new Window class.

Also changed AppContext syntax, now is has function-style member calls.